backend/requirements.txt
numpy
pandas
Pillow
scikit-learn == 1.2.0
flask
gunicorn